# ENH: _lib.doccer: Simplify and optimize indentation loop