@ilayn 2025-03-02
Just a question out of curiosity: Is it feasible to generate these docstrings once during packaging as a zlib compressed version and point to them in a static asset on demand instead of this massive undertaking to load up to memory as a mini encyclopedia?