@ev-br 2025-03-02
> Is it feasible to generate these docstrings once during packaging as a zlib compressed version and point to them in a static asset on demand instead of this massive undertaking to load up to memory as a mini encyclopedia?

Sure. There's no law of physics to say that it's impossible :-). Needs a champion though, and so far nobody had an itch bad enough to scratch. The current dance with filling the docstrings will need to stay as is for user subclasses.
The objects which we construct ourselves at import time, sure.

If somebody really wants to look into the scipy.stats import time, one other offender is likely a repeated `exec` in `_construct_argparser`: for N distributions we call exec N times.

All these potential improvements do no need to block this one though. Thanks @nickodell , all.