@nickodell 2025-03-03
> If somebody really wants to look into the scipy.stats import time, one other offender is likely a repeated `exec` in `_construct_argparser`: for N distributions we call exec N times.

That does look pretty suspicious. Some profiling suggests that `import scipy.stats` spends about 0.3% of its time executing the generated code inside `_attach_argparser_methods`, and about 0.4% of its time in `_construct_argparser` actually creating the code to execute.

Another piece of low hanging fruit seems to be docscrape: `import scipy.stats` spends 2.6% of its time inside docscrape.py, which is a library for parsing a docstring into its component parts, adding new sections, and turning that back into a string. This is used by e.g. the `_axis_nan_policy_factory` decorator.